refactor: split large components into smaller sub-components
Split AgentList.tsx into:
- AgentGridView.tsx - Grid display component
- AgentTableView.tsx - Table display component
- AgentListHeader.tsx - Filters and view toggle

Split DecisionFilters.tsx into:
- DateRangeFilter.tsx - Date range filter dropdown
- AgentFilter.tsx - Agent filter dropdown
- ConfidenceFilter.tsx - Confidence filter dropdown

Maintains full backwards compatibility by re-exporting
original exports from new component files.

Build passes, no logic changes - only code movement.

Closes #93
