+++ src/components/dashboard/agents/AgentListHeader.tsx
@@ -1,0 +1,165 @@
+'use client';
+
+import { LayoutGrid, List, Search, Filter, ChevronDown } from 'lucide-react';
+import { cn } from '@/lib/utils';
+import type { AgentStatus, AgentRole, ViewMode, SortField, SortOrder } from '@/types';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
+import type { Agent } from '@/types';
+
+interface AgentListHeaderProps {
+  searchQuery: string;
+  onSearchChange: (query: string) => void;
+  statusFilter: AgentStatus | 'all';
+  onStatusFilterChange: (status: AgentStatus | 'all') => void;
+  roleFilter: AgentRole | 'all';
+  onRoleFilterChange: (role: AgentRole | 'all') => void;
+  viewMode: ViewMode;
+  onViewModeChange: (mode: ViewMode) => void;
+  sortField: SortField;
+  onSortFieldChange: (field: SortField) => void;
+  sortOrder: SortOrder;
+  onSortOrderChange: (order: SortOrder) => void;
+}
+
+export function AgentListHeader({
+  searchQuery,
+  onSearchChange,
+  statusFilter,
+  onStatusFilterChange,
+  roleFilter,
+  onRoleFilterChange,
+  viewMode,
+  onViewModeChange,
+  sortField,
+  onSortFieldChange,
+  sortOrder,
+  onSortOrderChange,
+}: AgentListHeaderProps) {
+  return (
+    <div className="flex flex-col sm:flex-row gap-4 items-start sm:items-center justify-between">
+      <div className="flex flex-1 flex-wrap gap-3 items-center">
+        <div className="relative flex-1 min-w-[200px] max-w-sm">
+          <Search className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground" />
+          <Input
+            placeholder="Search agents..."
+            value={searchQuery}
+            onChange={(e) => onSearchChange(e.target.value)}
+            className="pl-9"
+          />
+        </div>
+
+        <Select value={statusFilter} onValueChange={(v) => onStatusFilterChange(v as AgentStatus | 'all')}>
+          <SelectTrigger className="w-[140px]">
+            <Filter className="h-4 w-4 mr-2" />
+            <SelectValue placeholder="Status" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="all">All Status</SelectItem>
+            <SelectItem value="active">Active</SelectItem>
+            <SelectItem value="idle">Idle</SelectItem>
+            <SelectItem value="paused">Paused</SelectItem>
+            <SelectItem value="error">Error</SelectItem>
+          </SelectContent>
+        </Select>
+
+        <Select value={roleFilter} onValueChange={(v) => onRoleFilterChange(v as AgentRole | 'all')}>
+          <SelectTrigger className="w-[140px]">
+            <SelectValue placeholder="Role" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="all">All Roles</SelectItem>
+            <SelectItem value="manager">Manager</SelectItem>
+            <SelectItem value="worker">Worker</SelectItem>
+            <SelectItem value="specialist">Specialist</SelectItem>
+            <SelectItem value="system">System</SelectItem>
+          </SelectContent>
+        </Select>
+
+        <Select value={sortField} onValueChange={(v) => onSortFieldChange(v as SortField)}>
+          <SelectTrigger className="w-[140px]">
+            <span className="text-muted-foreground mr-2">Sort:</span>
+            <SelectValue />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="name">Name</SelectItem>
+            <SelectItem value="status">Status</SelectItem>
+            <SelectItem value="last_active">Last Active</SelectItem>
+            <SelectItem value="tasks_completed">Tasks</SelectItem>
+          </SelectContent>
+        </Select>
+
+        <Button
+          variant="ghost"
+          size="icon"
+          onClick={() => onSortOrderChange(sortOrder === 'asc' ? 'desc' : 'asc')}
+        >
+          <ChevronDown className={cn('h-4 w-4 transition-transform', sortOrder === 'asc' && 'rotate-180')} />
+        </Button>
+      </div>
+
+      <div className="flex items-center gap-2 border rounded-lg p-1">
+        <Button
+          variant={viewMode === 'grid' ? 'secondary' : 'ghost'}
+          size="icon"
+          className="h-8 w-8"
+          onClick={() => onViewModeChange('grid')}
+        >
+          <LayoutGrid className="h-4 w-4" />
+        </Button>
+        <Button
+          variant={viewMode === 'list' ? 'secondary' : 'ghost'}
+          size="icon"
+          className="h-8 w-8"
+          onClick={() => onViewModeChange('list')}
+        >
+          <List className="h-4 w-4" />
+        </Button>
+      </div>
+    </div>
+  );
+}
+
+export function filterAndSortAgents(
+  agents: Agent[],
+  searchQuery: string,
+  statusFilter: AgentStatus | 'all',
+  roleFilter: AgentRole | 'all',
+  sortField: SortField,
+  sortOrder: SortOrder
+): Agent[] {
+  return agents
+    .filter((agent) => {
+      const matchesSearch = 
+        !searchQuery ||
+        agent.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
+        agent.description?.toLowerCase().includes(searchQuery.toLowerCase()) ||
+        agent.role.toLowerCase().includes(searchQuery.toLowerCase());
+
+      const matchesStatus = statusFilter === 'all' || agent.status === statusFilter;
+      const matchesRole = roleFilter === 'all' || agent.role === roleFilter;
+
+      return matchesSearch && matchesStatus && matchesRole;
+    })
+    .sort((a, b) => {
+      let comparison = 0;
+
+      switch (sortField) {
+        case 'name':
+          comparison = a.name.localeCompare(b.name);
+          break;
+        case 'status':
+          comparison = a.status.localeCompare(b.status);
+          break;
+        case 'last_active':
+          comparison = (a.last_active_at || '').localeCompare(b.last_active_at || '');
+          break;
+        case 'tasks_completed':
+          comparison = ((a.metadata?.tasks_completed as number) || 0) - ((b.metadata?.tasks_completed as number) || 0);
+          break;
+      }
+
+      return sortOrder === 'asc' ? comparison : -comparison;
+    });
+}
