+++ src/components/dashboard/decisions/DateRangeFilter.tsx
@@ -1,0 +1,61 @@
+'use client';
+
+import { Calendar, ChevronDown } from 'lucide-react';
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuCheckboxItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu';
+
+export type DateRange = 'all' | 'today' | 'week' | 'month';
+
+interface DateRangeFilterProps {
+  dateRange: DateRange;
+  onDateRangeChange: (range: DateRange) => void;
+}
+
+const DATE_RANGE_OPTIONS: { value: DateRange; label: string }[] = [
+  { value: 'all', label: 'All Time' },
+  { value: 'today', label: 'Today' },
+  { value: 'week', label: 'Last 7 Days' },
+  { value: 'month', label: 'Last 30 Days' },
+];
+
+export function DateRangeFilter({ dateRange, onDateRangeChange }: DateRangeFilterProps) {
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant="outline" className="gap-2">
+          <Calendar className="h-4 w-4" />
+          <span>Date</span>
+          {dateRange !== 'all' && (
+            <Badge variant="secondary" className="ml-1 h-5 px-1.5">
+              1
+            </Badge>
+          )}
+          <ChevronDown className="h-3 w-3 ml-1" />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="start" className="w-48">
+        <DropdownMenuLabel>Filter by Date</DropdownMenuLabel>
+        <DropdownMenuSeparator />
+        {DATE_RANGE_OPTIONS.map((option) => (
+          <DropdownMenuCheckboxItem
+            key={option.value}
+            checked={dateRange === option.value}
+            onCheckedChange={() => onDateRangeChange(option.value)}
+          >
+            {option.label}
+          </DropdownMenuCheckboxItem>
+        ))}
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
+
+export { DATE_RANGE_OPTIONS };
